fix: use local (old) DV chart plugin. DHIS2-7614
The classic DV app (and plugin) is going to be removed in 2.33, which
will cause this app to not load any charts.
After discussing about the possible solutions, due to time/effort
constraints, we decided to simply bundle the plugin and link it
internally.

(cherry picked from commit 175c8efc638e94110643bb3247603cacf213561f)